Profile backend fixed and battletag fixed
Issue with json decoding on profile page along with timeouts.  Added a hardcoded timeout for that page.

For battletag searching its possible for values to exist in the db that dont have a # in it, so for example, zero.  And it returned off the exact match function, but multiple, which was unexpected.  Put in checks for this

Also made sure the same account doesnt fill all of the alt battletag cookies